pyproject.toml
[project]
name = "status-api"
version = "0.0.0a"
description = "Self-Hosted REST-API to check on your Server's Status."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "psutil>=7.0.0",
    "py-cpuinfo>=9.0.0",
]
